+++ index.js
@@ -8,6 +8,7 @@ module.exports = {
     'no-chain': require('./rules/no-chain'),
     'no-extraneous-function-wrapping': require('./rules/no-extraneous-function-wrapping'),
     'no-single-composition': require('./rules/no-single-composition'),
+    'no-submodule-destructuring': require('./rules/no-submodule-destructuring'),
     'prefer-constant': require('./rules/prefer-constant'),
     'prefer-get': require('./rules/prefer-get'),
     'use-fp': require('./rules/use-fp')
@@ -28,6 +29,7 @@ module.exports = {
         'lodash-fp/no-chain': 'error',
         'lodash-fp/no-extraneous-function-wrapping': 'error',
         'lodash-fp/no-single-composition': 'error',
+        'lodash-fp/no-submodule-destructuring': 'error',
         'lodash-fp/prefer-constant': ['error', false],
         'lodash-fp/prefer-get': 'error',
         'lodash-fp/use-fp': 'error'